Change int64 to int in config

+++ internal/config/config.go
@@ -3,14 +3,14 @@ package config
 type Config struct {
 	GrpcServer struct {
 		Host string `yaml:"host"`
-		Port int64  `yaml:"port"`
+		Port int    `yaml:"port"`
 	} `yaml:"grpc_server"`
 	HttpServer struct {
 		Host string `yaml:"host"`
-		Port int64  `yaml:"port"`
+		Port int    `yaml:"port"`
 	} `yaml:"http_server"`
 	LseqDb struct {
 		Host string `yaml:"host"`
-		Port int64  `yaml:"port"`
+		Port int    `yaml:"port"`
 	} `yaml:"lseqdb"`
 }
